+++ inst/testScripts/BNR/00.params.R
@@ -1,22 +1,22 @@
 library("sansSouci")
 
-m <- 1e3
-B <- 1e3
-
-rho <- 0
-pi0 <- 0.9
-SNR <- 2
+m <- 1e2
+B <- 1e2
 
 alpha <- 0.25
-nbSimu <- 1e3
-typeOfSNR <- c("constant", "Pareto", "linear")[3]
+nbSimu <- 1e2
+typeOfSNR <- c("constant", "Pareto", "linear")[1]
 
-flavor <- c("equi", "Mein2006")[1]
+flavor <- c("equi", "Mein2006", "Toeplitz")[3]
 
 SNRs <- rev(c(0, 1, 2, 3, 4, 5))
 #SNRs <- "Pareto(2,2,1)"
 #SNRs <- c("Pareto(0,2,1)", "Pareto(1,2,1)", "Pareto(2,2,1)"); library("actuar");
-rhos <- c(0, 0.2, 0.4)
+if (flavor == "Toeplitz") {
+    deps <- c(-2, -1, -0.5, -0.2)  ## Toeplitz exponent
+} else {
+    deps <- c(0, 0.2, 0.4) ## correlation coefficient
+}
 pi0s <- c(0.8, 0.9, 0.99, 0.999)
 
 source("package/inst/testScripts/BNR/testStepDown.R")