+++ inst/testScripts/BNR/stepDownEqui+power.R
@@ -13,9 +13,9 @@ path <- file.path(resPath, sname, pname)
 path <- R.utils::Arguments$getWritablePath(path)
 
 for (SNR in SNRs) {
-    for (rho in rhos) {
+    for (dep in deps) {
         for (pi0 in pi0s) {
-            tags <- sprintf("pi0=%s,rho=%s,SNR=%s", pi0, rho, SNR)
+            tags <- sprintf("pi0=%s,dep=%s,SNR=%s", pi0, dep, SNR)
             filename <- sprintf("%s.rds", gsub("\\.", "_", tags))
             print(tags)
 
@@ -24,7 +24,7 @@ for (SNR in SNRs) {
                     print(bb)
                 }
 
-                rr <- testStepDown(m, rho, B, pi0, SNR, typeOfSNR, alpha, flavor=flavor)
+                rr <- testStepDown(m, dep, B, pi0, SNR, typeOfSNR, alpha, flavor=flavor)
             }, mc.cores=ncores)
 
             ## tidy? too many rows! (m*nbSimu>=1e6 per setting...)
@@ -49,7 +49,7 @@ for (SNR in SNRs) {
                 mat <- rbind(JFWER=eJR, Power=ePow, V0=eV0, S1=eS1)
                 names(dimnames(mat)) <- c("risk", "method")
                 dat <- reshape2::melt(mat)
-                dat <- cbind(pi0=pi0, rho=rho, SNR=SNR, dat)
+                dat <- cbind(pi0=pi0, dep=dep, SNR=SNR, dat)
             }
             pathname <- file.path(path, filename)
             saveRDS(dat, file=pathname)